Controller pageable, MySQL implementation

+++ backend/src/main/java/com/example/backend/models/dto/ClientDto.java
@@ -2,5 +2,5 @@ package com.example.backend.models.dto;
 
 import java.io.Serializable;
 
-public record ClientDto(Integer id, String username, String password) implements Serializable {
+public record ClientDto(Long id, String username, String password) implements Serializable {
 }
