+++ backend/src/main/java/com/example/backend/models/dto/ClientStockDto.java
@@ -3,6 +3,6 @@ package com.example.backend.models.dto;
 import java.io.Serializable;
 import java.math.BigDecimal;
 
-public record ClientStockDto(Integer id, Integer clientId, Integer stockId, Integer shares,
+public record ClientStockDto(Long id, Long clientId, Long stockId, Integer shares,
                              BigDecimal total) implements Serializable {
 }
