+++ backend/src/main/java/com/example/backend/models/dto/ClientVo.java
@@ -3,6 +3,6 @@ package com.example.backend.models.dto;
 import java.io.Serializable;
 import java.math.BigDecimal;
 
-public record ClientVo(Integer id, String username, BigDecimal balance,
+public record ClientVo(Long id, String username, BigDecimal balance,
                        BigDecimal virtualBalance) implements Serializable {
 }
