+++ backend/src/main/java/com/example/backend/models/dto/StockDto.java
@@ -4,6 +4,6 @@ import java.io.Serializable;
 import java.math.BigDecimal;
 import java.time.Instant;
 
-public record StockDto(Integer id, String symbol, String name, BigDecimal price,
+public record StockDto(Long id, String symbol, String name, BigDecimal price,
                        Instant modifiedAt) implements Serializable {
 }
