+++ backend/src/main/java/com/example/backend/models/dto/StockHistoryDto.java
@@ -4,6 +4,5 @@ import java.io.Serializable;
 import java.math.BigDecimal;
 import java.time.Instant;
 
-public record StockHistoryDto(Integer id, Integer stockId, BigDecimal price,
-                              Instant createdAt) implements Serializable {
+public record StockHistoryDto(Long id, Long stockId, BigDecimal price, Instant createdAt) implements Serializable {
 }
