+++ backend/src/main/java/com/example/backend/models/dto/TransactionDto.java
@@ -4,6 +4,6 @@ import java.io.Serializable;
 import java.math.BigDecimal;
 import java.time.Instant;
 
-public record TransactionDto(Integer id, Integer clientId, Integer stockId, Integer shares, BigDecimal total,
+public record TransactionDto(Long id, Long clientId, Long stockId, Integer shares, BigDecimal total,
                              Instant createdAt) implements Serializable {
 }
